requirements.txt
Django==5.0.1
django-ckeditor-5==0.2.12
django-otp==1.5.0
qrcode==7.4.2
django-admin-honeypot==1.2.0
django-ratelimit==4.1.0
django-cors-headers==4.3.1
pillow==10.2.0
bleach==6.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-csp==3.8
django-permissions-policy==4.20.0
python-magic==0.4.27
